materials/requirements.txt
flask-restplus==0.12.1 
Flask-SQLAlchemy==2.3.2 
pytest-flask==0.14.0 
SQLAlchemy==1.3.0
